feat: create common listSubstates provider method (#16)
After https://github.com/tari-project/tari-dan/pull/1060 and
https://github.com/tari-project/tari-snap/pull/19 the functionality for
substate listing is now the same for both wallet daemon and metamask
snap.

This PR creates a common `listSubstate` method in the `provider`
interface, with the corresponding implementation for each of the
providers (wallet daemon and snap). For the `tari_universe` provider the
method was left unimplemented, it would need to be done in a future PR.

For convenience, this PR also sets the tari `npm` depedencies using
`gitpkg` to always use the latest version of the bindings. This is meant
as a temporary solution until npm package deployment is integrated into
the tari-dan CI.