feat: monorepo conversion (#42)
Description
---

In order to support both browser and nodejs environments we need to
allow users to choose which packages to import.
The first step along that road is to structure the project as a monorepo
of multiple packages.
At this time all the packages are required for tarijs, however in the
future these requirements will be eliminated and instead passed as
parameters based on which provider package is installed.

How Has This Been Tested?
---
Building the project


What process can a PR reviewer use to test or verify this change?
---
after installing proto run 
`moon tarijs:build` then navigate to packages/tarijs directory. you
should have the build output in dist subdirectory. You should also run
the integration tests


<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

---------

Co-authored-by: Martins Erts <fp@tari.com>