Add docker build process (#55)
Description
---
Add a Dockerfile build process

Motivation and Context
---
Repeatable way of building the lib without cluttering the host machine.

How Has This Been Tested?
---
I have run the build in the Dockerfile.

What process can a PR reviewer use to test or verify this change?
---

Follow the instructions in the README.md
https://github.com/tari-project/tari.js/pull/55/files#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R34

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->