feat(package): add tarijs-all package (#58)
Description
---
Rename `@tari-project/tarijs` to `@tari-project/tarijs-all` to
distinguish monorepo from it's `tarijs` package.
Bump version to **0.5.3**.
Also some packages and tests structure refactor.

Motivation and Context
---
The `@tari-project/tarijs` package name was defined twice: in root and
`./packages/tarijs` folder. This caused confusion and was not clear what
this meant to be. So as clarification `@tari-project/tarijs` from the
package was renamed to `@tari-project/tarijs-all` to export all types
from other packages. Also cleanup of this package was done and tests
were adjusted and restructured.

How Has This Been Tested?
---
Manually.

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->