feat: add additional signer methods (#85)
Description
---

Adds the following methods to TariSigner:
- accountsList
- getAccountByAddress
- getWalletInfo

Motivation and Context
---

Flow editor needs new functionality.
First, it needs to determine the network, that the wallet connection is
bound to. It will be returned as part of `getWalletInfo` response.
Second, the editor has to list available accounts and get their details.
That is what `accountsList` and `getAccountByAddress` are for.

How Has This Been Tested?
---

Breaking Changes
---

- [x] None
- [ ] Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Describe the breaking change -->